Move studio registries to a more appropriate directory
