+++ packages/editor/src/components/element/ElementList.tsx
@@ -31,13 +31,13 @@ import { useTranslation } from 'react-i18next'
 
 import InputText from '@etherealengine/client-core/src/common/components/InputText'
 import { Component } from '@etherealengine/ecs/src/ComponentFunctions'
+import { ComponentShelfCategoriesState } from '@etherealengine/editor/src/services/ComponentShelfCategoriesState'
 import { getMutableState, getState, useHookstate } from '@etherealengine/hyperflux'
-import { ComponentShelfCategoriesState } from '@etherealengine/ui/src/components/editor/panels/Properties/elementList'
 import Icon from '@etherealengine/ui/src/primitives/mui/Icon'
 import Typography from '@etherealengine/ui/src/primitives/mui/Typography'
 import { ItemTypes } from '../../constants/AssetTypes'
-import { ComponentEditorsState } from '../../functions/ComponentEditors'
 import { EditorControlFunctions } from '../../functions/EditorControlFunctions'
+import { ComponentEditorsState } from '../../services/ComponentEditors'
 import { SelectionState } from '../../services/SelectionServices'
 import { usePopoverContextClose } from './PopoverContext'
