+++ packages/ui/src/components/editor/panels/Hierarchy/node/index.tsx
@@ -49,9 +49,9 @@ import { UUIDComponent } from '@etherealengine/ecs'
 import useUpload from '@etherealengine/editor/src/components/assets/useUpload'
 import { HeirarchyTreeNodeType } from '@etherealengine/editor/src/components/hierarchy/HeirarchyTreeWalker'
 import { ItemTypes, SupportedFileTypes } from '@etherealengine/editor/src/constants/AssetTypes'
-import { ComponentEditorsState } from '@etherealengine/editor/src/functions/ComponentEditors'
 import { EditorControlFunctions } from '@etherealengine/editor/src/functions/EditorControlFunctions'
 import { addMediaNode } from '@etherealengine/editor/src/functions/addMediaNode'
+import { ComponentEditorsState } from '@etherealengine/editor/src/services/ComponentEditors'
 import { SelectionState } from '@etherealengine/editor/src/services/SelectionServices'
 import { ResourcePendingComponent } from '@etherealengine/engine/src/gltf/ResourcePendingComponent'
 import { ErrorComponent } from '@etherealengine/engine/src/scene/components/ErrorComponent'
